+++ src/cli/run.rs
@@ -1,5 +1,6 @@
 use std::collections::{HashMap, HashSet, VecDeque};
-use std::path::PathBuf;
+use std::env;
+use std::path::{Path, PathBuf};
 use std::string::String;
 
 use clap::Parser;
@@ -38,6 +39,14 @@ pub struct Args {
     /// The path to 'pixi.toml'
     #[arg(long)]
     pub manifest_path: Option<PathBuf>,
+
+    /// Require pixi.lock is up-to-date
+    #[clap(long, conflicts_with = "frozen")]
+    pub locked: bool,
+
+    /// Don't check if pixi.lock is up-to-date, install as lockfile states
+    #[clap(long, conflicts_with = "locked")]
+    pub frozen: bool,
 }
 
 pub fn order_tasks(
@@ -81,6 +90,7 @@ pub fn order_tasks(
                 Execute {
                     cmd: CmdArgs::from(tasks),
                     depends_on: vec![],
+                    cwd: Some(env::current_dir().unwrap_or(project.root().to_path_buf())),
                 }
                 .into(),
                 Vec::new(),
@@ -141,25 +151,33 @@ pub async fn create_script(task: Task, args: Vec<String>) -> miette::Result<Sequ
     deno_task_shell::parser::parse(full_script.trim()).map_err(|e| miette!("{e}"))
 }
 
+/// Select a working directory based on a given path or the project.
+pub fn select_cwd(path: Option<&Path>, project: &Project) -> miette::Result<PathBuf> {
+    Ok(match path {
+        Some(cwd) if cwd.is_absolute() => cwd.to_path_buf(),
+        Some(cwd) => {
+            let abs_path = project.root().join(cwd);
+            if !abs_path.exists() {
+                miette::bail!("Can't find the 'cwd': '{}'", abs_path.display());
+            }
+            abs_path
+        }
+        None => project.root().to_path_buf(),
+    })
+}
 /// Executes the given command within the specified project and with the given environment.
 pub async fn execute_script(
     script: SequentialList,
-    project: &Project,
     command_env: &HashMap<String, String>,
+    cwd: &Path,
 ) -> miette::Result<i32> {
     // Execute the shell command
-    Ok(deno_task_shell::execute(
-        script,
-        command_env.clone(),
-        project.root(),
-        Default::default(),
-    )
-    .await)
+    Ok(deno_task_shell::execute(script, command_env.clone(), cwd, Default::default()).await)
 }
 
 pub async fn execute_script_with_output(
     script: SequentialList,
-    project: &Project,
+    cwd: &Path,
     command_env: &HashMap<String, String>,
     input: Option<&[u8]>,
 ) -> RunOutput {
@@ -170,7 +188,7 @@ pub async fn execute_script_with_output(
     drop(stdin_writer); // prevent a deadlock by dropping the writer
     let (stdout, stdout_handle) = get_output_writer_and_handle();
     let (stderr, stderr_handle) = get_output_writer_and_handle();
-    let state = ShellState::new(command_env.clone(), project.root(), Default::default());
+    let state = ShellState::new(command_env.clone(), cwd, Default::default());
     let code = execute_with_pipes(script, state, stdin, stdout, stderr).await;
     RunOutput {
         exit_code: code,
@@ -188,10 +206,11 @@ pub async fn execute(args: Args) -> miette::Result<()> {
     let mut ordered_commands = order_tasks(args.task, &project)?;
 
     // Get the environment to run the commands in.
-    let command_env = get_task_env(&project).await?;
+    let command_env = get_task_env(&project, args.locked, args.frozen).await?;
 
     // Execute the commands in the correct order
     while let Some((command, args)) = ordered_commands.pop_back() {
+        let cwd = select_cwd(command.working_directory(), &project)?;
         // Ignore CTRL+C
         // Specifically so that the child is responsible for its own signal handling
         // NOTE: one CTRL+C is registered it will always stay registered for the rest of the runtime of the program
@@ -200,7 +219,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
         let ctrl_c = tokio::spawn(async { while tokio::signal::ctrl_c().await.is_ok() {} });
         let script = create_script(command, args).await?;
         let status_code = tokio::select! {
-            code = execute_script(script, &project, &command_env) => code?,
+            code = execute_script(script, &command_env, &cwd) => code?,
             // This should never exit
             _ = ctrl_c => { unreachable!("Ctrl+C should not be triggered") }
         };
@@ -217,9 +236,13 @@ pub async fn execute(args: Args) -> miette::Result<()> {
 /// activation scripts from the environment and stores the environment variables it added, it adds
 /// environment variables set by the project and merges all of that with the system environment
 /// variables.
-pub async fn get_task_env(project: &Project) -> miette::Result<HashMap<String, String>> {
+pub async fn get_task_env(
+    project: &Project,
+    frozen: bool,
+    locked: bool,
+) -> miette::Result<HashMap<String, String>> {
     // Get the prefix which we can then activate.
-    let prefix = get_up_to_date_prefix(project).await?;
+    let prefix = get_up_to_date_prefix(project, frozen, locked).await?;
 
     // Get environment variables from the activation
     let activation_env = run_activation_async(project, prefix).await?;