+++ src/cli/shell.rs
@@ -1,4 +1,4 @@
-use crate::Project;
+use crate::{prompt, Project};
 use clap::Parser;
 use miette::IntoDiagnostic;
 use rattler_conda_types::Platform;
@@ -24,11 +24,20 @@ pub struct Args {
     /// The path to 'pixi.toml'
     #[arg(long)]
     manifest_path: Option<PathBuf>,
+
+    /// Require pixi.lock is up-to-date
+    #[clap(long, conflicts_with = "frozen")]
+    locked: bool,
+
+    /// Don't check if pixi.lock is up-to-date, install as lockfile states
+    #[clap(long, conflicts_with = "locked")]
+    frozen: bool,
 }
 
 fn start_powershell(
     pwsh: PowerShell,
     env: &HashMap<String, String>,
+    prompt: String,
 ) -> miette::Result<Option<i32>> {
     // create a tempfile for activation
     let mut temp_file = tempfile::Builder::new()
@@ -40,11 +49,13 @@ fn start_powershell(
     for (key, value) in env {
         shell_script.set_env_var(key, value);
     }
+    temp_file
+        .write_all(shell_script.contents.as_bytes())
+        .into_diagnostic()?;
 
-    let mut contents = shell_script.contents;
-    // TODO: build a better prompt
-    contents.push_str("\nfunction prompt {\"PS pixi> \"}");
-    temp_file.write_all(contents.as_bytes()).into_diagnostic()?;
+    // Write custom prompt to the env file
+    temp_file.write(prompt.as_bytes()).into_diagnostic()?;
+
     // close the file handle, but keep the path (needed for Windows)
     let temp_path = temp_file.into_temp_path();
 
@@ -59,7 +70,11 @@ fn start_powershell(
 }
 
 #[cfg(target_family = "windows")]
-fn start_cmdexe(cmdexe: CmdExe, env: &HashMap<String, String>) -> miette::Result<Option<i32>> {
+fn start_cmdexe(
+    cmdexe: CmdExe,
+    env: &HashMap<String, String>,
+    prompt: String,
+) -> miette::Result<Option<i32>> {
     // create a tempfile for activation
     let mut temp_file = tempfile::Builder::new()
         .suffix(".cmd")
@@ -74,6 +89,9 @@ fn start_cmdexe(cmdexe: CmdExe, env: &HashMap<String, String>) -> miette::Result
     temp_file
         .write_all(shell_script.contents.as_bytes())
         .into_diagnostic()?;
+
+    // Write custom prompt to the env file
+    temp_file.write(prompt.as_bytes()).into_diagnostic()?;
 
     let mut command = std::process::Command::new(cmdexe.executable());
     command.arg("/K");
@@ -93,6 +111,7 @@ async fn start_unix_shell<T: Shell + Copy>(
     shell: T,
     args: Vec<&str>,
     env: &HashMap<String, String>,
+    prompt: String,
 ) -> miette::Result<Option<i32>> {
     // create a tempfile for activation
     let mut temp_file = tempfile::Builder::new()
@@ -111,14 +130,26 @@ async fn start_unix_shell<T: Shell + Copy>(
         .write_all(shell_script.contents.as_bytes())
         .into_diagnostic()?;
 
+    // Write custom prompt to the env file
+    temp_file.write(prompt.as_bytes()).into_diagnostic()?;
+
     let mut command = std::process::Command::new(shell.executable());
     command.args(args);
 
-    let mut process = PtySession::new(command).into_diagnostic()?;
-    process
-        // Space added before `source` to automatically ignore it in history.
-        .send_line(&format!(" source {}", temp_file.path().display()))
+    // Space added before `source` to automatically ignore it in history.
+    let mut source_command = " ".to_string();
+    shell
+        .run_script(&mut source_command, temp_file.path())
         .into_diagnostic()?;
+
+    // Remove automatically added `\n`, if for some reason this fails, just ignore.
+    let source_command = source_command
+        .strip_suffix('\n')
+        .unwrap_or(source_command.as_str());
+
+    // Start process and send env activation to the shell.
+    let mut process = PtySession::new(command).into_diagnostic()?;
+    process.send_line(source_command).into_diagnostic()?;
 
     process.interact().into_diagnostic()
 }
@@ -166,9 +197,13 @@ async fn start_nu_shell(
 /// function as if the environment has been activated. This method runs the activation scripts from
 /// the environment and stores the environment variables it added, finally it adds environment
 /// variables from the project.
-pub async fn get_shell_env(project: &Project) -> miette::Result<HashMap<String, String>> {
+pub async fn get_shell_env(
+    project: &Project,
+    frozen: bool,
+    locked: bool,
+) -> miette::Result<HashMap<String, String>> {
     // Get the prefix which we can then activate.
-    let prefix = get_up_to_date_prefix(project).await?;
+    let prefix = get_up_to_date_prefix(project, frozen, locked).await?;
 
     // Get environment variables from the activation
     let activation_env = run_activation_async(project, prefix).await?;
@@ -192,7 +227,8 @@ pub async fn execute(args: Args) -> miette::Result<()> {
     let project = Project::load_or_else_discover(args.manifest_path.as_deref())?;
 
     // Get the environment variables we need to set activate the project in the shell.
-    let env = get_shell_env(&project).await?;
+    let env = get_shell_env(&project, args.frozen, args.locked).await?;
+    tracing::debug!("Pixi environment activation:\n{:?}", env);
 
     // Start the shell as the last part of the activation script based on the default shell.
     let interactive_shell: ShellEnum = ShellEnum::from_parent_process()
@@ -201,9 +237,13 @@ pub async fn execute(args: Args) -> miette::Result<()> {
 
     #[cfg(target_family = "windows")]
     let res = match interactive_shell {
-        ShellEnum::PowerShell(pwsh) => start_powershell(pwsh, &env),
-        ShellEnum::CmdExe(cmdexe) => start_cmdexe(cmdexe, &env),
         ShellEnum::NuShell(nushell) => start_nu_shell(nushell, &env).await,
+        ShellEnum::PowerShell(pwsh) => {
+            start_powershell(pwsh, &env, prompt::get_powershell_prompt(project.name()))
+        }
+        ShellEnum::CmdExe(cmdexe) => {
+            start_cmdexe(cmdexe, &env, prompt::get_cmd_prompt(project.name()))
+        }
         _ => {
             miette::bail!("Unsupported shell: {:?}", interactive_shell);
         }
@@ -211,12 +251,34 @@ pub async fn execute(args: Args) -> miette::Result<()> {
 
     #[cfg(target_family = "unix")]
     let res = match interactive_shell {
-        ShellEnum::PowerShell(pwsh) => start_powershell(pwsh, &env),
-        ShellEnum::Bash(bash) => start_unix_shell(bash, vec!["-l", "-i"], &env).await,
-        ShellEnum::Zsh(zsh) => start_unix_shell(zsh, vec!["-l", "-i"], &env).await,
-        ShellEnum::Fish(fish) => start_unix_shell(fish, vec![], &env).await,
-        ShellEnum::Xonsh(xonsh) => start_unix_shell(xonsh, vec![], &env).await,
         ShellEnum::NuShell(nushell) => start_nu_shell(nushell, &env).await,
+        ShellEnum::PowerShell(pwsh) => {
+            start_powershell(pwsh, &env, prompt::get_powershell_prompt(project.name()))
+        }
+        ShellEnum::Bash(bash) => {
+            start_unix_shell(
+                bash,
+                vec!["-l", "-i"],
+                &env,
+                prompt::get_bash_prompt(project.name()),
+            )
+            .await
+        }
+        ShellEnum::Zsh(zsh) => {
+            start_unix_shell(
+                zsh,
+                vec!["-l", "-i"],
+                &env,
+                prompt::get_zsh_prompt(project.name()),
+            )
+            .await
+        }
+        ShellEnum::Fish(fish) => {
+            start_unix_shell(fish, vec![], &env, prompt::get_fish_prompt(project.name())).await
+        }
+        ShellEnum::Xonsh(xonsh) => {
+            start_unix_shell(xonsh, vec![], &env, prompt::get_xonsh_prompt()).await
+        }
         _ => {
             miette::bail!("Unsupported shell: {:?}", interactive_shell)
         }