+++ tests/common/builders.rs
@@ -24,7 +24,7 @@
 //! ```
 
 use crate::common::IntoMatchSpec;
-use pixi::cli::{add, init, project, task};
+use pixi::cli::{add, init, install, project, task};
 use pixi::project::SpecType;
 use rattler_conda_types::Platform;
 use std::future::{Future, IntoFuture};
@@ -37,7 +37,7 @@ pub fn string_from_iter(iter: impl IntoIterator<Item = impl AsRef<str>>) -> Vec<
     iter.into_iter().map(|s| s.as_ref().to_string()).collect()
 }
 
-/// Contains the arguments to pass to `init::execute()`. Call `.await` to call the CLI execute
+/// Contains the arguments to pass to [`init::execute()`]. Call `.await` to call the CLI execute
 /// method and await the result at the same time.
 pub struct InitBuilder {
     pub args: init::Args,
@@ -71,7 +71,7 @@ impl IntoFuture for InitBuilder {
     }
 }
 
-/// Contains the arguments to pass to `add::execute()`. Call `.await` to call the CLI execute method
+/// Contains the arguments to pass to [`add::execute()`]. Call `.await` to call the CLI execute method
 /// and await the result at the same time.
 pub struct AddBuilder {
     pub args: add::Args,
@@ -109,6 +109,13 @@ impl AddBuilder {
         self
     }
 
+    /// Skip updating lockfile, this will only check if it can add a dependencies.
+    /// If it can add it will only add it to the manifest. Install will be skipped by default.
+    pub fn without_lockfile_update(mut self) -> Self {
+        self.args.no_lockfile_update = true;
+        self
+    }
+
     pub fn set_platforms(mut self, platforms: &[Platform]) -> Self {
         self.args.platform.extend(platforms.iter());
         self
@@ -139,6 +146,12 @@ impl TaskAddBuilder {
     /// Depends on these commands
     pub fn with_depends_on(mut self, depends: impl IntoIterator<Item = impl AsRef<str>>) -> Self {
         self.args.depends_on = Some(string_from_iter(depends));
+        self
+    }
+
+    /// With this working directory
+    pub fn with_cwd(mut self, cwd: PathBuf) -> Self {
+        self.args.cwd = Some(cwd);
         self
     }
 
@@ -199,5 +212,30 @@ impl IntoFuture for ProjectChannelAddBuilder {
             manifest_path: self.manifest_path,
             command: project::channel::Command::Add(self.args),
         }))
+    }
+}
+
+/// Contains the arguments to pass to [`install::execute()`]. Call `.await` to call the CLI execute method
+/// and await the result at the same time.
+pub struct InstallBuilder {
+    pub args: install::Args,
+}
+
+impl InstallBuilder {
+    pub fn with_locked(mut self) -> Self {
+        self.args.locked = true;
+        self
+    }
+    pub fn with_frozen(mut self) -> Self {
+        self.args.frozen = true;
+        self
+    }
+}
+
+impl IntoFuture for InstallBuilder {
+    type Output = miette::Result<()>;
+    type IntoFuture = Pin<Box<dyn Future<Output = Self::Output> + Send + 'static>>;
+    fn into_future(self) -> Self::IntoFuture {
+        Box::pin(install::execute(self.args))
     }
 }