+++ tests/common/mod.rs
@@ -4,7 +4,8 @@ pub mod builders;
 pub mod package_database;
 
 use crate::common::builders::{
-    AddBuilder, InitBuilder, ProjectChannelAddBuilder, TaskAddBuilder, TaskAliasBuilder,
+    AddBuilder, InitBuilder, InstallBuilder, ProjectChannelAddBuilder, TaskAddBuilder,
+    TaskAliasBuilder,
 };
 use pixi::cli::install::Args;
 use pixi::cli::run::{
@@ -13,8 +14,8 @@ use pixi::cli::run::{
 use pixi::cli::task::{AddArgs, AliasArgs};
 use pixi::cli::{add, init, project, run, task};
 use pixi::{consts, Project};
-use rattler_conda_types::conda_lock::CondaLock;
 use rattler_conda_types::{MatchSpec, PackageName, Platform, Version};
+use rattler_lock::CondaLock;
 
 use miette::IntoDiagnostic;
 use std::path::{Path, PathBuf};
@@ -66,7 +67,7 @@ impl LockFileExt for CondaLock {
     fn contains_package(&self, name: &PackageName) -> bool {
         self.package
             .iter()
-            .any(|locked_dep| locked_dep.name == *name)
+            .any(|locked_dep| locked_dep.name == name.as_normalized())
     }
 
     fn contains_matchspec(&self, matchspec: impl IntoMatchSpec) -> bool {
@@ -78,7 +79,7 @@ impl LockFileExt for CondaLock {
         self.package.iter().any(|locked_dep| {
             let package_version =
                 Version::from_str(&locked_dep.version).expect("could not parse version");
-            locked_dep.name == name && version.matches(&package_version)
+            locked_dep.name == name.as_normalized() && version.matches(&package_version)
         })
     }
 
@@ -96,7 +97,7 @@ impl LockFileExt for CondaLock {
         self.package.iter().any(|locked_dep| {
             let package_version =
                 Version::from_str(&locked_dep.version).expect("could not parse version");
-            locked_dep.name == name
+            locked_dep.name == name.as_normalized()
                 && version.matches(&package_version)
                 && locked_dep.platform == platform
         })
@@ -158,6 +159,7 @@ impl PixiControl {
                 specs: vec![spec.into()],
                 build: false,
                 no_install: true,
+                no_lockfile_update: false,
                 platform: Default::default(),
             },
         }
@@ -180,13 +182,17 @@ impl PixiControl {
         let mut tasks = order_tasks(args.task, &self.project().unwrap())?;
 
         let project = self.project().unwrap();
-        let task_env = get_task_env(&project).await.unwrap();
+        let task_env = get_task_env(&project, args.frozen, args.locked)
+            .await
+            .unwrap();
 
         let mut result = RunOutput::default();
         while let Some((command, args)) = tasks.pop_back() {
+            let cwd = run::select_cwd(command.working_directory(), &project)?;
             let script = create_script(command, args).await;
             if let Ok(script) = script {
-                let output = execute_script_with_output(script, &project, &task_env, None).await;
+                let output =
+                    execute_script_with_output(script, cwd.as_path(), &task_env, None).await;
                 result.stdout.push_str(&output.stdout);
                 result.stderr.push_str(&output.stderr);
                 result.exit_code = output.exit_code;
@@ -198,12 +204,15 @@ impl PixiControl {
         Ok(result)
     }
 
-    /// Create an installed environment. I.e a resolved and installed prefix
-    pub async fn install(&self) -> miette::Result<()> {
-        pixi::cli::install::execute(Args {
-            manifest_path: Some(self.manifest_path()),
-        })
-        .await
+    /// Returns a [`InstallBuilder`]. To execute the command and await the result call `.await` on the return value.
+    pub fn install(&self) -> InstallBuilder {
+        InstallBuilder {
+            args: Args {
+                manifest_path: Some(self.manifest_path()),
+                locked: false,
+                frozen: false,
+            },
+        }
     }
 
     /// Get the associated lock file
@@ -231,6 +240,7 @@ impl TasksControl<'_> {
                 commands: vec![],
                 depends_on: None,
                 platform,
+                cwd: None,
             },
         }
     }